working on adding a base protocol
