+++ CHANGES.txt
@@ -1,3 +1,6 @@
+3.11:
+    - drop binaries < 1.5.0
+    - create a base protocol
 3.10.1: hotfix for binary string
 3.10: add mc2 1.6.2 and 1.6.3 binaries, fix log name again
 3.9: remove MOTIONCOR2_PROGRAM, improve validation msg
