+++ motioncorr/__init__.py
@@ -32,15 +32,14 @@ import pyworkflow.utils as pwutils
 from .constants import *
 
 
-__version__ = '3.10.1'
+__version__ = '3.11'
 _references = ['Zheng2017']
 
 
 class Plugin(pwem.Plugin):
     _homeVar = MOTIONCOR2_HOME
     _pathVars = [MOTIONCOR2_CUDA_LIB]
-    _supportedVersions = ['1.4.0', '1.4.2', '1.4.4', '1.4.5', '1.4.7', '1.5.0',
-                          '1.6.2', '1.6.3']
+    _supportedVersions = ['1.5.0', '1.6.2', '1.6.3']
     _url = "https://github.com/scipion-em/scipion-em-motioncorr"
 
     @classmethod
@@ -52,7 +51,6 @@ class Plugin(pwem.Plugin):
         cudaVersion = cls.guessCudaVersion(MOTIONCOR2_CUDA_LIB)
         cls._defineVar(MOTIONCOR2_BIN, 'MotionCor2_1.6.3_Cuda%s%s_Feb18_2023' % (
             cudaVersion.major, cudaVersion.minor))
-
 
     @classmethod
     def getProgram(cls):
@@ -74,7 +72,7 @@ class Plugin(pwem.Plugin):
                         f"please verify {MOTIONCOR2_CUDA_LIB} variable."]
 
         except Exception as e:
-            return ["validateInstallation fails: %s" % e]
+            return [f"validateInstallation fails: {str(e)}"]
 
     @classmethod
     def versionGE(cls, version):
